demochandu/requirements.txt
Django==4.1.3
mysqlclient==2.1.1
djangorestframework==3.14.0
drf-yasg==1.21.4
markdown==3.4.1
django-filter==22.1
django-rest-swagger==2.2.0